02_productivity/shiny_app/requirements.txt
# Tung Chung PM2.5 Stats Explainer - Python dependencies
# Install with: pip install -r requirements.txt

shiny>=1.0.0
pandas>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
Jinja2>=3.0.0
